+++ source/B2B.Transactions/DataAccess/IUnitOfWork.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace B2B.Transactions
+namespace B2B.Transactions.DataAccess
 {
     /// <summary>
     /// Unit of work
